ADHOC fix (security): removing sensitive creds
removing the dependency on storing sensitive credentials on the server.
Instead those can be provided by the IAM EC2 Role, so that machine is allowed
to perform the required operations.
Sensitive credentials were in every cron email that was send from the server
in question, this is a big security hole, this fix is supposed to close it
in the case if server is using the iam role like described in the README now.

Also changing code for the compatibility with recent `aws cli` versions.
The implementation might be cleaner, so the improvements are welcome,
I guess the best way would be to switch to python and really do all
processing by parsing json resonses from AWS API. This can be considered in
future work.